243: remove CanisterTransport
