+++ src/candid_rpc.rs
@@ -1,52 +1,15 @@
 mod cketh_conversion;
 
-use crate::rpc_client::{EthRpcClient, MultiCallError, RpcTransport};
+use crate::rpc_client::{EthRpcClient, MultiCallError};
 use crate::{
-    accounting::get_http_request_cost,
     add_metric_entry,
     constants::ETH_GET_LOGS_MAX_BLOCKS,
-    http::http_request,
     providers::resolve_rpc_service,
-    types::{MetricRpcHost, MetricRpcMethod, ResolvedRpcService, RpcMethod},
+    types::{MetricRpcHost, ResolvedRpcService, RpcMethod},
 };
-use async_trait::async_trait;
 use candid::Nat;
 use ethers_core::{types::Transaction, utils::rlp};
-use evm_rpc_types::{
-    Hex, Hex32, MultiRpcResult, ProviderError, RpcApi, RpcError, RpcResult, RpcService,
-    ValidationError,
-};
-use ic_cdk::api::management_canister::http_request::{CanisterHttpRequestArgument, HttpResponse};
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-struct CanisterTransport;
-
-#[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
-#[cfg_attr(not(target_arch = "wasm32"), async_trait)]
-impl RpcTransport for CanisterTransport {
-    fn resolve_api(service: &RpcService) -> Result<RpcApi, ProviderError> {
-        Ok(resolve_rpc_service(service.clone())?.api())
-    }
-
-    async fn http_request(
-        service: &RpcService,
-        method: &str,
-        request: CanisterHttpRequestArgument,
-        effective_response_size_estimate: u64,
-    ) -> Result<HttpResponse, RpcError> {
-        let service = resolve_rpc_service(service.clone())?;
-        let cycles_cost = get_http_request_cost(
-            request
-                .body
-                .as_ref()
-                .map(|bytes| bytes.len() as u64)
-                .unwrap_or_default(),
-            effective_response_size_estimate,
-        );
-        let rpc_method = MetricRpcMethod(method.to_string());
-        http_request(rpc_method, service, request, cycles_cost).await
-    }
-}
+use evm_rpc_types::{Hex, Hex32, MultiRpcResult, RpcResult, ValidationError};
 
 fn process_result<T>(method: RpcMethod, result: Result<T, MultiCallError<T>>) -> MultiRpcResult<T> {
     match result {
@@ -79,7 +42,7 @@ fn process_result<T>(method: RpcMethod, result: Result<T, MultiCallError<T>>) ->
 }
 
 pub struct CandidRpcClient {
-    client: EthRpcClient<CanisterTransport>,
+    client: EthRpcClient,
 }
 
 impl CandidRpcClient {
@@ -205,7 +168,7 @@ fn get_transaction_hash(raw_signed_transaction_hex: &Hex) -> Option<Hex32> {
 mod test {
     use super::*;
     use crate::rpc_client::{MultiCallError, MultiCallResults};
-    use evm_rpc_types::RpcError;
+    use evm_rpc_types::{ProviderError, RpcError};
 
     #[test]
     fn test_process_result_mapping() {
