+++ src/rpc_client/mod.rs
@@ -6,16 +6,13 @@ use crate::rpc_client::eth_rpc::{
 use crate::rpc_client::numeric::TransactionCount;
 use crate::rpc_client::requests::GetTransactionCountParams;
 use crate::rpc_client::responses::TransactionReceipt;
-use async_trait::async_trait;
 use evm_rpc_types::{
     EthMainnetService, EthSepoliaService, HttpOutcallError, JsonRpcError, L2MainnetService,
-    ProviderError, RpcApi, RpcConfig, RpcError, RpcService, RpcServices,
+    ProviderError, RpcConfig, RpcError, RpcService, RpcServices,
 };
-use ic_cdk::api::management_canister::http_request::{CanisterHttpRequestArgument, HttpResponse};
 use serde::{de::DeserializeOwned, Serialize};
 use std::collections::BTreeMap;
 use std::fmt::Debug;
-use std::marker::PhantomData;
 
 pub mod checked_amount;
 pub(crate) mod eth_rpc;
@@ -35,40 +32,6 @@ macro_rules! log {
         // Print the message for convenience for local development (e.g. integration tests)
         println!("{}", &message);
     }}
-}
-
-#[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
-#[cfg_attr(not(target_arch = "wasm32"), async_trait)]
-pub trait RpcTransport: Debug {
-    fn resolve_api(provider: &RpcService) -> Result<RpcApi, ProviderError>;
-
-    async fn http_request(
-        provider: &RpcService,
-        method: &str,
-        request: CanisterHttpRequestArgument,
-        effective_size_estimate: u64,
-    ) -> Result<HttpResponse, RpcError>;
-}
-
-// Placeholder during refactoring
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct DefaultTransport;
-
-#[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
-#[cfg_attr(not(target_arch = "wasm32"), async_trait)]
-impl RpcTransport for DefaultTransport {
-    fn resolve_api(_provider: &RpcService) -> Result<RpcApi, ProviderError> {
-        unimplemented!()
-    }
-
-    async fn http_request(
-        _provider: &RpcService,
-        _method: &str,
-        _request: CanisterHttpRequestArgument,
-        _effective_size_estimate: u64,
-    ) -> Result<HttpResponse, RpcError> {
-        unimplemented!()
-    }
 }
 
 #[derive(Clone, Copy, Default, Debug, Eq, PartialEq)]
@@ -93,15 +56,14 @@ impl EthereumNetwork {
 }
 
 #[derive(Clone, Debug, PartialEq, Eq)]
-pub struct EthRpcClient<T: RpcTransport> {
+pub struct EthRpcClient {
     chain: EthereumNetwork,
     /// *Non-empty* list of providers to query.
     providers: Vec<RpcService>,
     config: RpcConfig,
-    phantom: PhantomData<T>,
 }
 
-impl<T: RpcTransport> EthRpcClient<T> {
+impl EthRpcClient {
     pub fn new(source: RpcServices, config: Option<RpcConfig>) -> Result<Self, ProviderError> {
         const DEFAULT_ETH_MAINNET_SERVICES: &[EthMainnetService] = &[
             EthMainnetService::Ankr,
@@ -174,7 +136,6 @@ impl<T: RpcTransport> EthRpcClient<T> {
             chain,
             providers,
             config: config.unwrap_or_default(),
-            phantom: PhantomData,
         })
     }
 
@@ -208,7 +169,7 @@ impl<T: RpcTransport> EthRpcClient<T> {
                 "[sequential_call_until_ok]: calling provider: {:?}",
                 provider
             );
-            let result = eth_rpc::call::<T, _, _>(
+            let result = eth_rpc::call::<_, _>(
                 provider,
                 method.clone(),
                 params.clone(),
@@ -254,7 +215,7 @@ impl<T: RpcTransport> EthRpcClient<T> {
             for provider in providers {
                 log!(DEBUG, "[parallel_call]: will call provider: {:?}", provider);
                 fut.push(async {
-                    eth_rpc::call::<T, _, _>(
+                    eth_rpc::call::<_, _>(
                         provider,
                         method.clone(),
                         params.clone(),
